chore(typescript): migrate connectNumericMenu
I didn't change the behaviour here at all, but there seems to be something either wrong in the helper types, or in its usage here.
