+++ src/types/connector.ts
@@ -3,6 +3,9 @@ import { Hits, InstantSearch } from './instantsearch';
 import { InsightsClient } from './insights';
 
 export type RendererOptions<TWidgetParams = unknown> = {
+  /**
+   * Original parameters for this widget, in case rendering depends on them.
+   */
   widgetParams: TWidgetParams;
   instantSearchInstance: InstantSearch;
   results?: SearchResults;
